RECOMMEND/hp_django/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-filter==22.1
djangorestframework==3.13.1
mysqlclient==2.1.1
numpy==1.23.3
pandas==1.4.4
Pillow==9.2.0
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
